feat: 대화 주제 조회 응답 DTO 재정의(#26)
